SM5 2023-Add Trustee Committee Treasurers
Update Memos for QMISK, TMEIT, ITK and the reception to have a person responsible for the committee's budget to be approved by SM.

For the reception, this is moved away from NÅGON to this new board member (mux) role. Splitting up the previous 3 mux roles into 1 budget responsible and 2 normal muxes.

Editorial changes were made to Memo for Reception part §2 Organization to make it more readable.

[SM#5 2023 Protocol](https://drive.google.com/file/d/1DZXxVFiWpFHqLGuprCFmDi-qO1WYdOh1/view)

[Proposition](https://docs.google.com/document/d/1gKDulRXm4NpKQMrL1FvphgzSEq87oTMnEBhMLxLPsmI/view)