SM1 2025-Update §1.10.1 Chapter Birthday
THS still classifies the chapter’s birthdate as 12 September 2000, the
day the IT-chapter was created. In our statutes, it says 4 December
2008, which is the date of the merger of the IT and ME chapters. Now as
we are going back to the IT chapter, it is appropriate to update the
bylaws so that the birthday of the chapter is the same as that of THS,
September 12, 2000.

[Proposition](https://docs.google.com/document/d/1ay4purcdbTjeX4aCEwnRSRjtuAj3qDlUdnHAufLtL2c/view)
[SM#5 2024
Protocol](https://drive.google.com/file/d/1XgNsyC4V0BXmR4DkHLyUU6prqMWiQ70H/view)
[SM#1 2025
Protocol](https://drive.google.com/file/d/1iz0_dA4c6tH8rZ8Y0Rui7uLD6hlAayY-/view)